Sparql Manager rename and pyright fixes
