+++ src/csvcubed/utils/csvdataset.py
@@ -34,12 +34,12 @@ from csvcubed.cli.inspect.inspectdatasetmanager import (
     get_standard_shape_unit_col_name_from_dsd,
 )
 from csvcubed.utils.sparql_handler.data_cube_state import DataCubeState
-from csvcubed.utils.sparql_handler.sparqlmanager import (
+from csvcubed.utils.sparql_handler.sparqlquerymanager import (
     CubeShape,
     select_single_unit_from_dsd,
 )
 from csvcubed.models.sparqlresults import QubeComponentResult
-from csvcubed.utils.sparql_handler.sparqlmanager import select_single_unit_from_dsd
+from csvcubed.utils.sparql_handler.sparqlquerymanager import select_single_unit_from_dsd
 
 
 def _materialise_unit_uri_for_row(
